🚚 Move applications to macos folder
