+++ macos/applications.go
@@ -1,15 +1,15 @@
-package applications
+package macos
 
 import (
 	"os/exec"
 	"strings"
 
 	"github.com/Matt-Gleich/nuke/output"
-	"github.com/Matt-Gleich/statuser"
+	"github.com/Matt-Gleich/statuser/v2"
 )
 
-// Get ... Get a list of all the open application
-func Get() []string {
+// GetApplications ... Get a list of all the open application
+func GetApplications() []string {
 	out, err := exec.Command("osascript", "-e", `tell application "System Events" to get name of (processes where background only is false)`).Output()
 	if err != nil {
 		err := exec.Command("osascript", "-e", `tell application "System Events" to activate`).Run()
@@ -27,8 +27,8 @@ func Get() []string {
 	return cleanApplications
 }
 
-// Quit ... Quit an application
-func Quit(name string) {
+// QuitApp ... Quit an application
+func QuitApp(name string) {
 	cleanedName := strings.ReplaceAll(name, " ", "\\ ")
 	err := exec.Command("pkill", "-x", cleanedName).Run()
 	if err != nil {
